extension: make compatible with libnm
libnm-glib has been deprecated in favor of libnm for ages. GNOME Shell
seems to be among the last bits to be ported. This is now being done [1]:

[1] https://bugzilla.gnome.org/show_bug.cgi?id=789811

Unfortunately, due to GLib limitations, libnm-glib.so and libnm.so can't
be loaded at the same time and the attempt to do so results in a nasty
crash. The extensions thus needs to be updated to libnm.

There's a bug [2] with recent gjs that makes this crash. Hopefully fixed by
GNOME 2.28, with or before the libnm-based Shell.

[2] https://gitlab.gnome.org/GNOME/gjs/issues/9